backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
crewai>=0.108.0
crewai-tools>=0.38.0
firecrawl-py>=1.0.0
docling>=2.90.0
supabase>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
